Update analysis-services.md
I found a bit unnatural sentence in Japanese language and would like to fix it. This proposal is supposed to be addressed as a `Linguistic Suggestion` from my point of view. It would be appreciated if you would check this proposal. Please let me know if you have any concerns.

Best regards,