Update a sentence for description
I found a bit unnatural sentence in Japanese language and would like to fix it. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you would check this proposal. Please let me know if you have any concerns.

This request is in regard to:

- [Run the SQL Server 2017 container image with Docker](https://docs.microsoft.com/en-us/sql/linux/quickstart-install-connect-docker)
- [Docker を使用した SQL Server 2017 コンテナー イメージを実行します。](https://docs.microsoft.com/ja-jp/sql/linux/quickstart-install-connect-docker)

Best regards,