Update uninstall-an-existing-instance-of-sql-server-setup.md
Hi Team, I found a bit unnatural sentence in Japanese language, and I would like to fix it. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you would check this proposal. Please let me know if you have any concerns.

This request is in regard to:

- [Uninstall an Existing Instance of SQL Server (Setup)](https://docs.microsoft.com/en-us/sql/sql-server/install/uninstall-an-existing-instance-of-sql-server-setup)
- [SQL Server の既存のインスタンスのアンインストール (セットアップ)](https://docs.microsoft.com/ja-jp/sql/sql-server/install/uninstall-an-existing-instance-of-sql-server-setup)

Best regards,