Update name-sos-short.md
Hi Team, I found a translation which is obviously wrong in Japanese language. I am not sure that the included-file is aligned to your acceptance criteria. But I would like to open this pull-request because I believe the tool name should be described as `SQL Operations Studio`, not `SQL 操作 Studio`. It looks unnatural in Japanese language. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you could take a look at this pull-request. Please let me know if you have any concerns.

Best regards,