Update appliesto-ss-asdb-asdw-pdw-md.md
Hi Team,

I found an unnecessary punctuation. This PR suggests getting rid of that. And this PR is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you could take a look at this pull-request. Please let me know if you have any concerns.

This request is in regard to:

- https://github.com/MicrosoftDocs/sql-docs.ja-jp/blob/live/docs/includes/appliesto-ss-asdb-asdw-pdw-md.md

Best regards,