Update wide-world-importers-oltp-database-catalog.md
Hi Team,

I found a translation that is obviously wrong in Japanese language, and I would like to fix it. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you could take a look at this pull-request. Please let me know if you have any concerns.

This request is in regard to:

- https://docs.microsoft.com/ja-jp/sql/samples/wide-world-importers-oltp-database-catalog
- https://docs.microsoft.com/en-us/sql/samples/wide-world-importers-oltp-database-catalog

Best regards,